pyproject.toml
[tool.poetry]
name = "deploy-function-action"
version = "0.1.0"
description = ""
authors = ["Magnus Moan <magnus.moan@cognite.com>"]

[tool.black]
line-length = 120
target_version = ['py38']
include = '\.py$'

[tool.isort]
line_length=120                # corresponds to -w  flag
multi_line_output=3            # corresponds to -m  flag
include_trailing_comma=true    # corresponds to -tc flag
skip_glob = '^((?!py$).)*$'    # this makes sort all Python files

[tool.poetry.dependencies]
python = "^3.8"
cognite-sdk-experimental = "0.19.0"
pydantic = "1.6.1"
python-crontab = "2.5.1"
"ruamel.yaml" = "0.16.10"

[tool.poetry.dev-dependencies]
bandit = "1.6.2"
black = "20.8b1"
isort = "5.4.2"
mypy = "0.782"
pytest = "6.0.1"
pytest-cov = "^2.10.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
